feat(login) front-end login - modify folder structure [continues #163519099]
