+++ src/__tests__/components/App.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import App from '../../components/App';
+
+describe('<App />', () => {
+  test('Should render the APP', () => {
+    const app = shallow(<App />);
+    expect(app).toMatchSnapshot();
+  });
+});
