+++ src/__tests__/components/Home.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Home from '../../components/Home/Home';
+
+describe('<Home />', () => {
+  test('Should render the <Home />', () => {
+    const wrapper = shallow(<Home />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
