+++ src/__tests__/components/common/Input.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Input from '../../../components/common/Input/Input';
+
+describe('<Input />', () => {
+  test('Should render the <Input />', () => {
+    const handleInput = () => true;
+    const wrapper = shallow(<Input onChange={handleInput} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
