+++ src/__tests__/store/store.test.js
@@ -1,0 +1,7 @@
+import store from '../../redux/store';
+
+describe('redux-store', () => {
+  it('should return the initial state', () => {
+    expect(store.getState()).toBeDefined();
+  });
+});
