--- src/components/__tests__/actions/userActions.test.js
@@ -1,13 +1,0 @@
-import { setUserProfile } from '../../../redux/actions/userActions';
-import { signupUser } from '../../../__mocks__/dummyData';
-import { SET_PROFILE } from '../../../redux/actions-types/userTypes';
-
-describe('userActions', () => {
-  test("should update user's data", () => {
-    const expectedAction = {
-      type: SET_PROFILE,
-      payload: signupUser,
-    };
-    expect(setUserProfile(signupUser)).toEqual(expectedAction);
-  });
-});
